==> 3.1_WordGrepHimani.ipynb <==
Himani Parikh - 1322085

CSCI 636 - Big Data Analytics

Hands on 3.1 : WordCount - Grep Command

***Question:***

Download the text to Alice's Adventures in Wonderland from
http://www.gutenberg.org/files/11/11-0.txt
(If it redirects you to a page with a welcome popup, click on the "Plain Text UTF-
8" option on that page or just download the attachment below) and run
wordcount on it. This can be done by using hadoop commands. How many times
does the word Cheshire occur? (Do not include the word 'Cheshire with an
apostrophe. The string -->'Cheshire<-- does not count)

Hint: grep is in the same jar file as the wordcount.

*Step* 1: Installing Hadoop italicized text

#Download Hadoop installation package

!wget https://downloads.apache.org/hadoop/common/hadoop-3.3.6/hadoop-3.3.6.tar.gz

#Use the tar command with the -x flag to extract, -z to uncompress,
#-v for verbose output, and -f to specify that we’re extracting from a file

!tar -xzvf hadoop-3.3.6.tar.gz

#copying the hadoop file to user/local
!cp -r hadoop-3.3.6/ /usr/local/

#To find the default Java path
!readlink -f /usr/bin/java | sed "s:bin/java::"

#set java home variable
import os
os.environ["JAVA_HOME"] = "/usr/lib/jvm/java-11-openjdk-amd64"

#Running Hadoop
!/usr/local/hadoop-3.3.6/bin/hadoop

# Access files in Google Drive
from google.colab import drive
drive.mount('/content/drive')

# example	MapReduce	programs
!/usr/local/hadoop-3.3.6/bin/hadoop jar /usr/local/hadoop-3.3.6/share/hadoop/mapreduce/hadoop-mapreduce-examples-3.3.6.jar

# Wordcount	MapReduce	programs
!/usr/local/hadoop-3.3.6/bin/hadoop jar /usr/local/hadoop-3.3.6/share/hadoop/mapreduce/hadoop-mapreduce-examples-3.3.6.jar wordcount

# Verify	input	file	exists
!ls '/content/drive/My Drive/Colab Notebooks/BigDataColabHimani/Data/'

# Run	WordGrep.
!/usr/local/hadoop-3.3.6/bin/hadoop jar \
/usr/local/hadoop-3.3.6/share/hadoop/mapreduce/hadoop-mapreduce-examples-3.3.6.jar \
grep \
'/content/drive/My Drive/Colab Notebooks/BigDataColabHimani/Data/alice.txt' \
'/content/drive/My Drive/Colab Notebooks/BigDataColabHimani/out_grep' \
"[^']Cheshire"

# See	WordGrep	output	directory
!ls '/content/drive/My Drive/Colab Notebooks/BigDataColabHimani/out_grep'

# View the WordGrep results.
!more '/content/drive/My Drive/Colab Notebooks/BigDataColabHimani/out_grep/part-r-00000'
