==> 7.Spark-WordCountHimani.ipynb <==
Himani Parikh - 1322085

CSCI 636 - Big Data Analytics

Hands on 7.1 : WordCount with Hadoop Spark

***1. Install pyspark***

!pip install pyspark

***2. Initiate	the	spark	context***

from pyspark import SparkContext
from pyspark.sql import SparkSession

spark = SparkSession\
        .builder\
        .master('local[*]')\
        .getOrCreate()
sc = spark.sparkContext

***Mount	your	Google	Drive	and	change	the	working	directory***


from google.colab import drive
drive.mount('/content/drive')

# change the directory to big-data-colab folder
%cd /content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh

#pwd
!ls

***Read	Shakespeare	text	from	Google	drive.***

lines = sc.textFile('/content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh/Data/words.txt')

lines.count()

***Split	each	line	into	words.***

words = lines.flatMap(lambda line : line.split(" "))

***Assign initial count value to each word.***

tuples = words.map(lambda word : (word,1))

***Sum all word count values***

counts = tuples.reduceByKey(lambda a,b: (a+b))

***Prepare	the	output	directory.***

!rm -r 'SparkOutPut'

***Write word counts to text file in HDFS***

counts.coalesce(1).saveAsTextFile('/content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh/SparkOutput')

***Check the files in the output directory***

!ls SparkOutput

***View the wordcount results***

!head -20 SparkOutput/part-00000
