==> 8.Spark-SQLHimani.ipynb <==
Himani Parikh - 1322085

CSCI 636 - Big Data Analytics

Hands on 8.1 : Spark SQL

***Install pyspark***

!pip install pyspark

***Initiate the spark context***

from pyspark import SparkContext
from pyspark.sql import SparkSession

spark = SparkSession\
        .builder\
        .master('local[*]')\
        .getOrCreate()
sc = spark.sparkContext

***Mount Google drive***

from google.colab import drive
drive.mount('/content/drive')

# change the directory to big-data-colab folder
%cd /content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh

#pwd
!ls

***Connect to SQL Table***

from pyspark.sql import SQLContext
sqlsc = SQLContext(sc)

df = sqlsc.read.load('/content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh/SqlDataSet/game-clicks.csv',
                     format='com.databricks.spark.csv',
                     header='true',
                     inferSchema='true')

***View Spark DataFrame schema and count rows***

df.printSchema()

df.count()

***View contents of DataFrame.***

df.show(5)

***Filter columns in DataFrame.***

df.select("userid", "teamlevel").show(5)

***Filter rows based on criteria.***

df.filter(df["teamlevel"] > 1).select("userid", "teamlevel").show(5)

***Group by a column and count.***

df.groupBy("ishit").count().show()

***Calculate average and sum***

from pyspark.sql.functions import*

df.select(mean('ishit'), sum('ishit')).show()

***Join two DataFrames.***

df2 = sqlsc.read.load('/content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh/SqlDataSet/ad-clicks.csv',
                     format='com.databricks.spark.csv',
                     header='true',
                     inferSchema='true')

df2.printSchema()

merge = df.join(df2, 'userid')

***Let's view the schema of merge:***

merge.printSchema()

merge.show(5)
