==> 8.Spark-StreamingHimani.ipynb <==
Himani Parikh - 1322085

CSCI 636 - Big Data Analytics

Hands on 8.2 : Spark Streaming

***1. Install pyspark***

!pip install pyspark

import multiprocessing

cores = multiprocessing.cpu_count()
cores

from pyspark import SparkContext
from pyspark.sql import SparkSession

spark = SparkSession\
        .builder\
        .master('local[*]')\
        .getOrCreate()
sc = spark.sparkContext

import re
def parse(line):
  match = re.search("Dm=(\d+)", line)
  if match:
    val = match.group(1)
    return [int(val)]
  return []

***Import and create streaming context.***

from pyspark.streaming import StreamingContext
ssc = StreamingContext(sc,1)

Create Dstream of waether data

lines =  ssc.socketTextStream("rtd.hpwren.ucsd.edu", 12028)

Read measurement

vals = lines.flatMap(parse)

***create sliding of window data***

window = vals.window(10,5)

***Define and call analysis function:***

def stats(rdd):
  print(rdd.collect())
  if rdd.count() > 0:
    print("max = {}, min ={}".format(rdd.max(), rdd.min()))

window.foreachRDD(lambda rdd: stats(rdd))

***Start the stream processing***

ssc.start()
ssc.awaitTermination()
