==> 10.Spark-ClassificationHimani.ipynb <==
Himani Parikh - 1322085

CSCI 636 - Big Data Analytics

Hands on 10.1 : Spark-Classification using decision decision tree(low-humidity days vs. non-low- humidity days at 9am)

***Install pyspark***

!pip install pyspark

from pyspark import SparkContext
from pyspark.sql import SparkSession

spark = SparkSession\
        .builder\
        .master('local[*]')\
        .getOrCreate()
sc = spark.sparkContext

***Mount Google Drive***

from google.colab import drive
drive.mount('/content/drive')

# change the directory to big-data-colab folder
%cd /content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh

#pwd
!ls

***Load classes and data***

from pyspark.sql import SQLContext
from pyspark.sql import DataFrameNaFunctions

from pyspark.ml import Pipeline
from pyspark.ml.classification import DecisionTreeClassifier
from pyspark.ml.feature import Binarizer
from pyspark.ml.feature import VectorAssembler, StringIndexer, VectorIndexer

sqlsc = SQLContext(sc)

df = sqlsc.read.load('/content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh/WeatherDataset/daily_weather.csv',
                     format='com.databricks.spark.csv',
                     header='true',
                     inferSchema='true')

df.columns

featureColumns = ['air_pressure_9am',
 'air_temp_9am',
 'avg_wind_direction_9am',
 'avg_wind_speed_9am',
 'max_wind_direction_9am',
 'max_wind_speed_9am',
 'rain_accumulation_9am',
 'rain_duration_9am']

***Drop unused and missing data***

df = df.drop('number')

df = df.na.drop()
df.count(), len(df.columns)

***Create categorical variable***

binarizer = Binarizer(threshold=24.99999, inputCol='relative_humidity_3pm', outputCol='label')
binarizerDF = binarizer.transform(df)

binarizerDF.select('relative_humidity_3pm', 'label').show(4)

***Aggregate features***

assembler = VectorAssembler(inputCols=featureColumns, outputCol="features")
assembled = assembler.transform(binarizerDF)

***Split training and test data***

(trainingData, testData) = assembled.randomSplit([0.8, 0.2], seed=13234)

trainingData.count(), testData.count()

***Create and train decision tree***

dt = DecisionTreeClassifier(labelCol="label", featuresCol="features", maxDepth=5,
                       minInstancesPerNode=20, impurity="gini")

pipeline = Pipeline(stages=[dt])
model = pipeline.fit(trainingData)

predictions = model.transform(testData)

predictions.select("prediction", "label").show(10)

***Save predicitions to CSV***

predictions.select("prediction", "label").write.save(path="/content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh/Prediction/predictions.csv",
                                                     format='com.databricks.spark.csv',
                                                     header='true')
