==> 11.2.Spark-Classifier-evaluationHimani-Quiz.ipynb <==
Himani Parikh - 1322085

CSCI 636 - Big Data Analytics

Hands on 11.2 Quiz : Spark Evaluation of Decision Tree (Accuracy and Confusion Matrix)

***Install pyspark***

!pip install pyspark

from pyspark import SparkContext
from pyspark.sql import SparkSession

spark = SparkSession\
        .builder\
        .master('local[*]')\
        .getOrCreate()
sc = spark.sparkContext

***Mount Google Drive***

from google.colab import drive
drive.mount('/content/drive')

# change the directory to big-data-colab folder
%cd /content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh

#pwd
!ls

***Load predictions***

from pyspark.sql import SQLContext

from pyspark.ml.evaluation import MulticlassClassificationEvaluator
from pyspark.mllib.evaluation import MulticlassMetrics

sqlsc = SQLContext(sc)

predictions = sqlsc.read.load('/content/drive/My Drive/Colab Notebooks/BigDataColabHimaniParikh/Prediction/predictions.csv',
                     format='com.databricks.spark.csv',
                     header='true',
                     inferSchema='true')

***Compute accuracy***

***Question 2:***


---


To print out the accuracy as a percentage, use the following code:

A.
print ("Accuracy = %.2g" % (accuracy * 100)) ✅

B.
print ("Accuracy = %100g" % (accuracy))

C.
print ("Accuracy = %100.2g" % (accuracy))

D.
None of the above


---


Answer: A ✅

evaluator = MulticlassClassificationEvaluator(labelCol="label", predictionCol="prediction", metricName="accuracy")
accuracy = evaluator.evaluate(predictions)
print("Accuracy = %.2g%%" % (accuracy * 100))


***Question 1:***

---


To get the error rate for the decision tree model, use the following code:

A.
print ("Error = %g " % (1.0 - accuracy)) ✅

B.
evaluator = MuticlassClassificationEvaluator(
labelCol="label",
predictionCol="prediction",
metricName="error")

C.
error = evaluator.evaluate(1 - predictions)

D.
None of the abov


---
Answer: A ✅


print("Error = %g" % (1.0 - accuracy))

***Display confusion matrix***

predictions.rdd.take(2)

predictions.rdd.map(tuple).take(2)

metrics = MulticlassMetrics(predictions.rdd.map(tuple))

***Question 3:***

---



In the last line of code, the confusion matrix is printed out. If the “transpose()” is
removed, the confusion matrix will be displayed as:

A.
array([[87., 14.],
[26., 83.]])

B.
array([[83., 26.],
[14., 87.]])

C.
array([[83., 87.],
[14., 26.]])

D.
None of the above ✅


---


***Answer: D*** ✅

metrics.confusionMatrix().toArray()
